fix(docker): deleted oci runtime integratin
The OCI provider was installed to isolate the Irie compiler output from
the rest of the system. Tests have shown that we can achieved the same
level of isolation (if not more) by combining apparmor and capabilities.

As such I removed all definition of the OCI runtime in the app.